crea profesor

+++ src/main/java/cl/awakelab/miprimerspring/service/serviceimpl/AlumnoServiceImpl.java
@@ -1,14 +1,12 @@
 package cl.awakelab.miprimerspring.service.serviceimpl;
 
 import cl.awakelab.miprimerspring.com.example.cl.proyecto.entity.entity.Alumno;
-import cl.awakelab.miprimerspring.com.example.cl.proyecto.entity.entity.Curso;
 import cl.awakelab.miprimerspring.com.example.cl.proyecto.entity.repository.IAlumnoRepository;
 import cl.awakelab.miprimerspring.com.example.cl.proyecto.entity.repository.ICursoRepository;
 import cl.awakelab.miprimerspring.service.IAlumnoService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
 import java.util.List;
 
 @Service("AlumnoServiceImpl")
@@ -17,50 +15,46 @@ public class AlumnoServiceImpl implements IAlumnoService {
     @Autowired
     IAlumnoRepository objAlumnoRepo;
 
-    @Autowired
-    ICursoRepository objCursoRepo;
+    //@Autowired//
+   // ICursoRepository objCursoRepo;//
 
     @Override
     public Alumno crearAlumno(Alumno alumnoCreado) {
-        Alumno nuevoAlumno = new Alumno();
-        Curso cursoAsignado = new Curso();
-        cursoAsignado = objCursoRepo.findById(alumnoCreado.getCurso().getId()).orElse(null);
-        alumnoCreado.setCurso(cursoAsignado);
-        nuevoAlumno = objAlumnoRepo.save(alumnoCreado);
+        Alumno nuevoAlumno =  objAlumnoRepo.save(alumnoCreado);
         return nuevoAlumno;
-
     }
 
     @Override
-    public Alumno actualizarAlumno(int id, Alumno alumno) {
-        Alumno alumnoActualizado = new Alumno();
-        alumnoActualizado = objAlumnoRepo.findById(alumno.getId()).orElse(null);
-        alumnoActualizado.setNombres(alumno.getNombres());
-        alumnoActualizado.setApellido1(alumno.getApellido1());
-        alumnoActualizado.setApellido2(alumno.getApellido2());
-        alumnoActualizado.setCurso(alumno.getCurso());
-        return alumnoActualizado;
+    public Alumno actualizarAlumno(Alumno alumno) {
+
+        return objAlumnoRepo.save(alumno);
+    }
+
+    @Override
+    public Alumno actualizarAlumno(int id, Alumno alumnoActualizado) {
+        return null;
     }
 
     @Override
     public List<Alumno> listarAlumnos() {
-        List <Alumno> listaAMostrarAlumnos =new ArrayList<>();
-        listaAMostrarAlumnos = objAlumnoRepo.findAll();
+        List <Alumno> listaAMostrarAlumnos = objAlumnoRepo.findAll();
         return listaAMostrarAlumnos;
     }
 
     @Override
     public Alumno listarAlumnosId(int id) {
+
         return objAlumnoRepo.findById(id).orElse(null);
     }
 
     @Override
-    public boolean eliminarAlumno(int id) {
-        objAlumnoRepo.deleteById(id);
-        return true;
+    public void eliminarAlumno(int id) {
+        if (objAlumnoRepo.existsById(id)) {
+            objAlumnoRepo.deleteById(id);
+        } else {
+            System.out.println("El alumno con ID " + id + " no existe.");
+        }
     }
-
-
