+++ src/main/java/cl/awakelab/miprimerspring/service/serviceimpl/ProfesorServiceImpl.java
@@ -6,7 +6,6 @@ import cl.awakelab.miprimerspring.service.IProfesorService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
 import java.util.List;
 @Service ("ProfesorServiceImpl")
 public class ProfesorServiceImpl implements IProfesorService {
@@ -16,39 +15,41 @@ public class ProfesorServiceImpl implements IProfesorService {
 
    @Override
     public Profesor crearProfesor(Profesor profesorCreado) {
-        Profesor nuevoProfesor = new Profesor();
-        nuevoProfesor= objProfesorRepo.save (profesorCreado);
+        Profesor nuevoProfesor= objProfesorRepo.save (profesorCreado);
        return nuevoProfesor;
 
     }
 
     @Override
     public Profesor actualizarProfesor(Profesor profesor) {
-       Profesor profesorActualizado = new Profesor();
-       profesorActualizado = objProfesorRepo.findById(profesor.getId()).orElse(null);
-       profesorActualizado.setNombres(profesor.getNombres());;
-       profesorActualizado.setApellido1(profesor.getApellido1());
-       profesorActualizado.setApellido2(profesor.getApellido2());
-       profesorActualizado.setListaCursos(profesor.getListaCursos());
-        return profesorActualizado;
+
+        return objProfesorRepo.save(profesor);
+    }
+
+    @Override
+    public Profesor actualizarProfesor(int id, Profesor profesorActualizado) {
+        return null;
     }
 
 
     @Override
     public List<Profesor> listarProfesores() {
-       List <Profesor> listaAMostrarProfesores =new ArrayList<>();
-       listaAMostrarProfesores = objProfesorRepo.findAll();
+       List <Profesor> listaAMostrarProfesores = objProfesorRepo.findAll();
         return listaAMostrarProfesores;
     }
 
     @Override
     public Profesor listarProfesoresId(int id) {
-        return objProfesorRepo.findById(id).orElse(null);
+
+       return objProfesorRepo.findById(id).orElse(null);
     }
 
     @Override
-    public boolean eliminarProfesor(int id) {
-       objProfesorRepo.deleteById(id);
-       return true;
+    public void eliminarProfesor(int id) {
+        if (objProfesorRepo.existsById(id)) {
+            objProfesorRepo.deleteById(id);
+        } else {
+            System.out.println("El profesor con ID " + id + " no existe.");
+        }
     }
 }
